fix: detect LISTEN death, drain backlogs, wire visibility-risk detection into watch_reclaims
- _watch_reclaims_pg: detect LISTEN-connection loss via asyncpg's
  termination listener + is_closed() instead of an except clause that
  could only ever fire on pool-side errors (asyncpg silently clears
  notification listeners on termination — nothing ever raises into the
  consume loop). Pool errors from poll_reclaim_events /
  _catch_up_after_notify now propagate honestly instead of being
  misdiagnosed as LISTEN loss, which degraded the watcher to poll-only
  for ~300s. Owned connections reconnect every
  _RECONNECT_POLL_INTERVAL poll iterations; caller-owned connections
  fall back to permanent polling (documented limitation).
- Drain full batches immediately (no poll_timeout between pages) in all
  watch_reclaims loops; a >100-event backlog previously drained at
  ~100 events per poll_timeout.
- Wire check_reclaim_visibility_delay_risk into both watch transports
  on a 60s cadence with a loud watch_reclaims-visibility-delay-at-risk
  warning, so a violated visibility-delay assumption is loud by default
  rather than opt-in; probe failures are logged and swallowed.
- Sweep 1 + isolate_self: an exhausted reclaim with an in-flight cancel
  lands on 'cancelled' (honest terminal label) instead of 'crashed';
  isolate_self mirrors _SWEEP_1_SQL branch-for-branch (cancel-state
  reset, clock_timestamp()) to preserve the isolate==sweep invariant
  asserted by test_leader_property.
- check_reclaim_visibility_risk: cast EXTRACT to float8 (asyncpg
  otherwise returns Decimal, which json.dumps rejects).
- Tests: fake-conn transport state-machine tests (reconnect,
  caller-owned fallback, pool-error propagation, full-batch drain),
  visibility-probe tests, cursor-on-empty-table and pruning-interaction
  tests, exactly-one-pg_notify-per-sweep test, a cancel-reset
  regression that reproduces the re-cancellation retry loop via
  poll_cancel_flags, and a real-asyncpg reconnect assertion in the
  LISTEN-kill chaos test. Each behavior fix verified to fail on the
  pre-fix code.
- Docs: watch_reclaims/poll_reclaim_events docstrings state the exact
  lost-event condition up front plus cursor/duplicate/pruning/shutdown/
  backpressure semantics; architecture.md documents default-on
  detection and adds a worked fan-out-completion example.
- testing/fixtures.py: pin dragonfly proactor_threads/maxmemory so
  redis fixtures start on high-core hosts instead of hanging.

+++ src/taskq/backend/_sql_templates.py
@@ -554,8 +554,12 @@ LIMIT $2""",
 -- actual miss: this cannot see whether that transaction will insert a
 -- job_events row at all, only that it has held the table open unusually
 -- long. Excludes this query's own backend.
+-- The ::float8 cast matters: EXTRACT returns numeric, which asyncpg
+-- hands back as Decimal — but LongRunningJobEventsWriter.xact_age_seconds
+-- is a float, and Decimal is not JSON-serializable for the monitoring
+-- loop this diagnostic feeds.
 SELECT a.pid, a.xact_start,
-       EXTRACT(EPOCH FROM (clock_timestamp() - a.xact_start)) AS xact_age_seconds
+       EXTRACT(EPOCH FROM (clock_timestamp() - a.xact_start))::float8 AS xact_age_seconds
 FROM pg_locks l
 JOIN pg_stat_activity a ON a.pid = l.pid
 WHERE l.relation = '"{s}".job_events'::regclass
